REF: move only-used-once mixins to resample (#41058)
